Adjusted Ewald force calculation
Note: Not yet completely happy with force implementation. Will not relax a random TiO2 structure in P1 in 100 iteration. Not sure whether this is a relaxation problem or a force problem. Also, Ewald equations I was using seemed to produce the negative of the force. Checked measurement manually (i.e. displace atom, calculate energy, approximate derivative), was within 1% of that result. So, I *think* this works but am not 100% positive.